Refactor ActuatorFAST
Create a inheritence level to store common code
between actuator lines and disks
